Sharded compactor  (#115)
* vertical block split

* Updated prometheus to v0.40.1


Fixed pkg


Fixed internal


Updated weaveworks/common


Fixed tests


Fixed some querier tests and commented out log spam


Fixed querier test


Reverted tls integration test


Fixed TestChunkQueryable


Implemented pushdownSeriesIterator AtT fn


Fixed cloudtrace import


Ran go mod tidy


trigger tests


* Ran go mod tidy


* merge


* Fixed binary reader and GetRef


* Renaming command line flag name

* Static compactor sharding

* tmp

* Fix tests

* Fix tests

* Fix logs

* Change grouping key

* Delete blocks after everything is compacted

* Emit a single log after compaction

* Do not retry successful tasks

* Change compaction key

* Use a numerical shard ID

* Shard only unsharded blocks

* Properly set sharding group

* Update UI and add duplicate detection

* Fix group key

* Upload shard 0 at the end

* Fix react lint

* Upload blocks in parallel

* Improve deletions

* Fixes duplicate deletions

* Reduce excess GCS calls

* Parallelize compactions

Currently blocks from Receivers and from Compactors are grouped
under the same compaction group. Since vertical compactions have
priority, this can sometimes cause them to block the horizontal
(time-wise) compaction backlog.

This commit adds the source as part of the group key. As a result
blocks coming from receivers and blocks uploaded by compactors
will be treated as separate groups and will be compacted in parallel.

* Disable vertical compaction for downsampled blocks

* Skip vertical compaction for multiple downsampled blocks

* Merge shopify main (#108)

* Merge shopify main

* Update mimir/prometheus

* Fix relabel.Process

* Fixed GenerateTestHistograms

* Fixed lint

---------


* Vendor mimir (#114)

* Tried to move mimir to internal, gave up for now

* Added mimir internal

* Fixed mimir internal

* Mimir internal

* Removed mimir import

* More fixed and mimir internal

* Lint and import fixes

* Fixed lint

* Fixed lint

* Excluded mimir internal from lint

* Fixed imports

* Fixed copyrights for mimir

* Exclude mimir from doc check

* Added todos remove mimir from white noise cleanup

* Sorted imports

* Make assets

* Synced and fixe dedupmimir

* Fixed block test

* Another fix for block_test.go

* Ignore mimir-prometheus in codeql

---------

+++ pkg/compact/planner_test.go
@@ -21,6 +21,7 @@ import (
 	"github.com/thanos-io/objstore"
 
 	"github.com/efficientgo/core/testutil"
+
 	"github.com/thanos-io/thanos/pkg/block"
 	"github.com/thanos-io/thanos/pkg/block/metadata"
 )
@@ -72,15 +73,17 @@ func TestPlanners_Plan_Compatibility(t *testing.T) {
 	}
 
 	// This mimics our default ExponentialBlockRanges with min block size equals to 20.
+	// TODO(rabenhorst) Test with sharded compactor.
 	tsdbComp, err := tsdb.NewLeveledCompactor(context.Background(), nil, nil, ranges, nil, nil)
 	testutil.Ok(t, err)
 	tsdbPlanner := &tsdbPlannerAdapter{comp: tsdbComp}
 	tsdbBasedPlanner := NewTSDBBasedPlanner(log.NewNopLogger(), ranges)
 
 	for _, c := range []struct {
-		name     string
-		metas    []*metadata.Meta
-		expected []CompactionTask
+		name              string
+		metas             []*metadata.Meta
+		expected          []CompactionTask
+		skipCompatibility bool
 	}{
 		{
 			name: "Outside range",
@@ -381,6 +384,24 @@ func TestPlanners_Plan_Compatibility(t *testing.T) {
 				},
 			},
 		},
+		{
+			name: "Downsampled overlapping blocks",
+			metas: []*metadata.Meta{
+				{
+					BlockMeta: tsdb.BlockMeta{Version: 1, ULID: ulid.MustNew(1, nil), MinTime: 0, MaxTime: 30},
+					Thanos:    metadata.Thanos{Downsample: metadata.ThanosDownsample{Resolution: 300000}},
+				},
+				{
+					BlockMeta: tsdb.BlockMeta{Version: 1, ULID: ulid.MustNew(2, nil), MinTime: 0, MaxTime: 60},
+					Thanos:    metadata.Thanos{Downsample: metadata.ThanosDownsample{Resolution: 300000}},
+				},
+				{
+					BlockMeta: tsdb.BlockMeta{Version: 1, ULID: ulid.MustNew(2, nil), MinTime: 0, MaxTime: 20},
+					Thanos:    metadata.Thanos{Downsample: metadata.ThanosDownsample{Resolution: 300000}},
+				},
+			},
+			skipCompatibility: true,
+		},
 	} {
 		t.Run(c.name, func(t *testing.T) {
 			for _, e := range c.expected {
@@ -396,6 +417,9 @@ func TestPlanners_Plan_Compatibility(t *testing.T) {
 
 			// For compatibility.
 			t.Run("tsdbPlannerAdapter", func(t *testing.T) {
+				if c.skipCompatibility {
+					return
+				}
 				dir, err := os.MkdirTemp("", "test-compact")
 				testutil.Ok(t, err)
 				defer func() { testutil.Ok(t, os.RemoveAll(dir)) }()
@@ -443,8 +467,8 @@ func TestRangeWithFailedCompactionWontGetSelected(t *testing.T) {
 		540,
 		1620,
 	}
-
 	// This mimics our default ExponentialBlockRanges with min block size equals to 20.
+	// TODO(rabenhorst) Test with sharded compactor.
 	tsdbComp, err := tsdb.NewLeveledCompactor(context.Background(), nil, nil, ranges, nil, nil)
 	testutil.Ok(t, err)
 	tsdbPlanner := &tsdbPlannerAdapter{comp: tsdbComp}
